Working sample ui code
